Shoot me (Alternatively: Creating Multiple Cards)
Added a banner pip for the king and queen names.
If in demo mode, the name for the playable cards does not show up to make the cards look more like the real life version.
Also decided to tackle creating multiples of each type of card in demo mode to create a full, game complient, deck. So far the playable cards all create the right number of themselves. (E.g. 4 knight cards, 3 dragon cards, ect.)
This only happens in demoMode. Not demo mode will only show one of each type of card.